+++ Changelog.md
@@ -45,3 +45,4 @@
 - [Updated some quest description to inform the player about charging items in storage units](https://github.com/TheStaticVoid/StaTech-Industry/issues/411)
 - [Add compressor recipe for netherite nuggets -> netherite ingot](https://github.com/TheStaticVoid/StaTech-Industry/issues/418)
 - [Add assembler recipe for breaking down transformers](https://github.com/TheStaticVoid/StaTech-Industry/issues/419)
+- [Nerfed TechReborn UU-Matter recipes for Uranium and Iridium](https://github.com/TheStaticVoid/StaTech-Industry/issues/413)